exercise updated

+++ smb_ws/src/smb_highlevel_controller/include/smb_highlevel_controller/SmbHighlevelController.hpp
@@ -49,7 +49,7 @@ private:
 
 	bool arrived;
 
-	int proportional;
+	float proportional;
 
 	int queue_size;
